requirements.txt
flask==2.0.1
sqlalchemy==1.4.46
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
python-dotenv==0.19.0
pandas==1.3.3
gunicorn==20.1.0
pymysql==1.0.2
werkzeug==2.0.3